refactor: type wgpu texture identity APIs

+++ examples/01-renderers/dear_app_wgpu_textures.rs
@@ -31,11 +31,7 @@ impl TexDemoState {
         let tex_w: u32 = 128;
         let tex_h: u32 = 128;
         let mut img_tex = dear_imgui_rs::texture::TextureData::new();
-        img_tex.create(
-            dear_imgui_rs::texture::TextureFormat::RGBA32,
-            tex_w as i32,
-            tex_h as i32,
-        );
+        img_tex.create(dear_imgui_rs::texture::TextureFormat::RGBA32, tex_w, tex_h);
 
         // Seed with a gradient for first frame
         let mut pixels = vec![0u8; (tex_w * tex_h * 4) as usize];
@@ -94,11 +90,7 @@ impl TexDemoState {
                     let (w, h) = rgba.dimensions();
                     let data = rgba.into_raw();
                     let mut t = dear_imgui_rs::texture::TextureData::new();
-                    t.create(
-                        dear_imgui_rs::texture::TextureFormat::RGBA32,
-                        w as i32,
-                        h as i32,
-                    );
+                    t.create(dear_imgui_rs::texture::TextureFormat::RGBA32, w, h);
                     t.set_data(&data);
                     println!("Loaded demo image from {:?} ({}x{})", path, w, h);
                     Some(t)
